Allow inputs with placeholders to shrink to value
If a AutosizeInput has a placeholder, the normal behaviour is to never
let the width go below that of the placeholder. This adds the ability to
supply a prop - `shrinkToValue` - which allows the input to autosize as
soon as something is entered in the input.

Also updates the example page with a relevant placeholder +
shrinkToValue example.